Restructure code and remove failing hack
This patch restructures the activities report gadget and removes
a failing hack that prevented the form from displaying when clicked.

Huge thanks to @Krinkle for helping me uncover the underlying issue
and fixing the code again.

Previously, the code worked by accident but meta-wiki has now changed
and the site loads same modules as the WADP making the wait time of
the gadget less which broke some things especially loading translation
strings.

Change-Id: I92ae67dbc4c63887c3714011ab5811404219e1bc