OOC: Tag talk page notification by level
Extract logic for posting to talk pages into a method on its own so
it's easier to pass in arguments by callers.

Also, pass in notice level of talk page notification in the generator.

At level 5, let M&E bypass be 'Yes' for all affiliates not only Chaps
or ThOrgs.

Change-Id: I3eacff444c666d9cbfd8716eb41f539f1c3f44ba