OrgInfoForm: Enable M&E staff only to input the affiliate code
This removes the step where by M&E staff will need to input the group
code in the machine readable list. Instead, what the M&E staff will do
now is just input the group code and name in the from and the system
will populate the MRL with relevant data needed to do autosuggest when
typing in the group name.

Change-Id: Ic41f682601be1dde89552155ee48252d91727500