Compute compliance status even in a new year 2021 -> 2022
This patch fixes the problem of loosing track of the OOC status/reporting
due date once the year ends and crosses over to a new year. We should
keep track of the reporting due date in the data base and compute based
on that.

When affiliate submits a report and resets the OOC level, also reset the
reporting due date.

Change-Id: Ic90a95a898c49b3cbe95d41c108f02ff10389f2b