frontend: Correctly mark user groups that are legal entities
This patch fixes a bug that made all user groups that are legal
entries to not go out of compliance when the reporting due date
is passed because the logic for computing the distance between
the latest activity report and the latest financial report was
flawed.

The actual distance between both reports should be less than 2
years instead of "greater than" 1 year.

Change-Id: I917f0003360a3b4d8cba5c75b996a0c62b848e64