OutOfComplianceChecker: Adjust timezone issue and sweep inconsistencies
Due to the time section in the reporting due date timestamp and client
requests timezone mismatch, the reporting due date was offset by about
12hrs+ which made sweeps not work in a timely manner.

This patch fixes the timezone issue partially but we still need a full
solution to consider the timezone of the server running the code instead
of client side zones sent via user agents.

In addition, we refactored the code that resets the reporting due date
of the affiliate so that we use the reporting due date of the affiliate
instead of the fiscal year which is incorrect.

Change-Id: Id2dee4d987e9db116939b1b5b22de4b46084ebd5